refactor(server): extract device/remote admin config flows (#3962 4.2a PR5 §4e)
Split the local config setters + edit-session flow (setDeviceConfig,
setLoRaConfig, setNetworkConfig, setChannelConfig, setPositionConfig,
setMQTTConfig, setNeighborInfoConfig, setDisplayConfig, setPowerConfig,
setTelemetryConfig, setGenericModuleConfig, setNodeOwner,
beginEditSettings/commitEditSettings, buildDeviceConfigFromActual) into
DeviceAdminService, and the remote-fetch flows (requestRemoteConfig,
requestRemoteChannel, requestRemoteOwner, requestRemoteDeviceMetadata,
requestRemoteSessionPasskey, sendRebootCommand, sendSetTimeCommand,
requestAllModuleConfigs, requestModuleConfig, refreshModuleConfigs,
resetModuleConfigCache) into a sibling RemoteAdminService.

Split rationale: combined the two groups exceed the ~800-line guideline,
and the remote-fetch flows have a materially different hazard shape —
remoteNodeConfigs/remoteNodeChannels/remoteNodeOwners/
remoteNodeDeviceMetadata/pendingModuleConfigRequests are written by
processAdminMessage (protobuf dispatch, out of scope) on packet receipt
and read/cleared by these outbound methods, the same split-brain shape
PR4's adminTransactionService documents for pendingAdminAcks.

buildDeviceConfigFromActual turned out to be cleanly separable (every
private field/method it touched already had, or gained, a narrow public
accessor) and moved along with the local setters; its dead
calculateLoRaFrequency private wrapper (now unused, only call site moved)
was deleted in favor of calling the util directly.

Both services are constructor-injected with `this` (import-cycle-safe,
`import type` only) per the established pattern from PR2-4's extractions.
None of the moved methods touch channel-vs-DM routing, so key-repair
invariant I6 is preserved by non-involvement.

meshtasticManager.ts: 14,102 -> 13,175 lines.
